Ability to change buffer size on the fly, faster SPI

+++ src/utility/socket.h
@@ -21,8 +21,8 @@ extern uint16_t sendto(SOCKET s, const uint8_t * buf, uint16_t len, uint8_t * ad
 extern uint16_t recvfrom(SOCKET s, uint8_t * buf, uint16_t len, uint8_t * addr, uint16_t *port); // Receive data (UDP/IP RAW)
 extern void flush(SOCKET s); // Wait for transmission to complete
 
-extern void setTXBuffer(SOCKET s, uint8_t buffer_size);
-extern void setRXBuffer(SOCKET s, uint8_t buffer_size);
+extern void setTXBuffer(SOCKET s, uint16_t buffer_size);
+extern void setRXBuffer(SOCKET s, uint16_t buffer_size);
 
 extern uint16_t igmpsend(SOCKET s, const uint8_t * buf, uint16_t len);
